backend/requirements.txt
aiodns==3.6.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiomultiprocess==0.9.1
aiosignal==1.4.0
alembic==1.16.5
amazon.ion==0.9.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.3
asn1crypto==1.5.1
asteval==1.0.6
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.9
awscli==1.44.58
bc-detect-secrets==1.5.45
bc-jsonpath-ng==1.6.1
bc-python-hcl2==0.4.3
bcrypt==4.2.0
beartype==0.22.9
beautifulsoup4==4.14.3
bitarray==3.7.1
black==26.3.1
boolean.py==5.0
boto3==1.42.68
botocore==1.42.68
CacheControl==0.14.4
cached-property==2.0.1
cachetools==5.5.2
cbor2==5.8.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
checkov==3.2.501
ckzg==2.1.1
click==8.2.1
click-option-group==0.5.9
cloudsplaining==0.7.0
colorama==0.4.6
ConfigArgParse==1.7.1
construct==2.10.68
construct-typing==0.5.6
contextlib2==21.6.0
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
cyclonedx-python-lib==7.6.2
cytoolz==1.0.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dockerfile-parse==2.0.1
docutils==0.19
dpath==2.1.3
elementpath==5.0.4
email_validator==2.2.0
emergent-plugins @ file:///wheelhouse/emergent_plugins-1.0.0-py3-none-any.whl#sha256=c0f704b1f3e22b0bfa65048252154d4524b82184a05034acfd677391e1d02c6e
emergentintegrations==0.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.121.3
ffmpeg-python==0.2.0
filelock==3.20.3
flake8==7.3.0
fonttools==4.61.0
fpdf2==2.8.6
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.30.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.3.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio-ffmpeg==0.6.0
importlib_metadata==7.2.1
iniconfig==2.1.0
ionhash==1.2.1
IPTCInfo3==2.1.4
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jq==1.10.0
jsonalias==0.1.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
junit-xml==1.9
kiwisolver==1.4.9
lark==1.3.1
license-expression==30.4.4
litellm==1.75.5.post1
lxml==6.0.1
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mdurl==0.1.2
motor==3.7.1
msgpack==1.1.2
multidict==6.6.4
mutagen==1.47.0
mypy==1.17.1
mypy_extensions==1.1.0
networkx==2.6.3
numpy==2.3.2
oauthlib==3.3.1
openai==1.99.9
orjson==3.11.7
packageurl-python==0.13.4
packaging==23.2
pandas==2.3.1
parsimonious==0.10.0
pathspec==1.0.4
paypalrestsdk==1.13.3
pillow==12.1.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
playwright==1.58.0
pluggy==1.6.0
ply==3.11
policy_sentry==0.13.2
prettytable==3.17.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.6
psutil==7.0.0
psycopg2-binary==2.9.10
py-serializable==1.1.2
py-solc-x==2.0.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycares==4.11.0
pycep-parser==0.5.1
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.1
pyexiv2==2.15.4
pyflakes==3.4.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.12.1
pymongo==4.15.4
PyNaCl==1.6.2
pyOpenSSL==26.0.0
pyparsing==3.2.3
pyqldb==3.2.4
pytest==8.4.1
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.22
python-xmp-toolkit==2.0.2
pytokens==0.4.1
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
qrcode==8.2
rdflib==7.6.0
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.21.0
rfc3986==1.5.0
rich==14.1.0
rlp==4.1.0
rpds-py==0.27.0
rsa==4.7.2
rustworkx==0.17.1
s3transfer==0.16.0
s5cmd==0.2.0
schedule==1.2.2
schema==0.7.5
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
solana==0.30.2
solders==0.18.1
sortedcontainers==2.4.0
soupsieve==2.8.3
spdx-tools==0.8.3
SQLAlchemy==2.0.43
starlette==0.50.0
stripe==12.4.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==2.3.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
typer==0.16.0
types-cachetools==4.2.10
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unidiff==0.7.5
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.3
uvicorn==0.25.0
watchfiles==1.1.0
wcwidth==0.6.0
web3==7.13.0
webauthn==1.11.1
websocket-client==1.9.0
websockets==15.0.1
xmlschema==4.1.0
xmltodict==1.0.2
yarl==1.20.1
zipp==3.23.0
